% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxygoodconfirm.R
\name{proxyBadConfirm}
\alias{proxyBadConfirm}
\title{confirm bad proxy}
\usage{
proxyBadConfirm(proxy_bad, .proxy)
}
\arguments{
\item{proxy_bad}{todo}

\item{.proxy}{todo}
}
\value{
todo
}
\description{
confirm bad proxy
}
