% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxygoodconfirm.R
\name{proxyGoodConfirm}
\alias{proxyGoodConfirm}
\title{confirm good proxy}
\usage{
proxyGoodConfirm(proxy_good, .proxy)
}
\arguments{
\item{proxy_good}{list with ip and port properties}

\item{.proxy}{list of list with ip and port properties, times}
}
\value{
list of list with `ip`, `port` and `times` properties
}
\description{
confirm good proxy
}
\details{
.proxy must be list class, if no elments in it, the return times will
 be 1, otherwise, the time value will be added 1
}
\examples{
\dontrun{
.proxy <- list(list(ip = '1.1.1.1', port = 111, times = 1),
               list(ip = '2.2.2.2', port = 222, times = 2))
 proxy_good <- list(ip = '1.1.1.1', port = 111)
.proxy <- proxyGoodConfirm(proxy_good, .proxy)
.proxy
}
}
