% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue2.R
\name{Rogue}
\alias{Rogue}
\title{Object to make http request}
\description{
Make http request with random proxy and useragent, try until got the response in
specific max iter times.
}
\section{Usage}{

\preformatted{
rogue <- Rogue$new(proxy = proxy, useragent = useragent, iter_max = 10,
                   is_record = TRUE, is_random = FALSE, is_quite = FALSE)

rogue$proxy_add(proxy, delete = FALSE)
rogue$proxy_show()

rogue$get(...)
rogue$post(...)
rogue$put(...)
rogue$delete(...)

rogue$proxy_show()
}
}

\section{Arguments}{

\describe{
  \item{proxy}{proxy list, with \code{ip} and \code{port} elements at least, the
    object initialization process will check the list, \code{times} element will
    be append if it does not exist, and the value is 1}
  \item{useragent}{multiple user agent character vector}
  \item{iter_max}{the max times to try send the http request, it should be not
    bigger than the length of proxy list}
  \item{is_record}{if record the history of connecting to proxy, the result will
    be used in select better proxy}
  \item{is_random}{if select proxy based on the probability, which is the successful
    connecting times for each proxy}
  \item{is_quite}{if output the error message when try to connect proxy}
  \item{delete}{if delete old and worse proxy when add new ones}
  \item{...}{arguments send to function of \code{httr} package}
}
}

\section{Details}{


\code{Rogue$new()} creates a new Rogue object
\code{rogue$proxy_add()} add new proxy for requesting
\code{rogue$proxy_show()} show the proxy connecting history in \code{tibble} format
\code{rogue$get()} send \code{get} request to server
\code{rogue$post()} send \code{post} request to server
\code{rogue$put()} send \code{put} request to server
\code{rogue$delete()} send \code{delete} request to server
}

