% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{useragent}
\alias{useragent}
\title{user agent strings}
\format{a character vector without names
\describe{
  user agent string
}}
\source{
somewhere on web
}
\usage{
useragent
}
\description{
a dataset containing 485 user agent strings
}
\keyword{datasets}
