% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogue.R
\name{ROGUE}
\alias{ROGUE}
\title{rogue function}
\usage{
ROGUE(.f = c("GET", "POST", "PUT", "DELETE"), ..., .proxy = NULL,
  .user_agent = NULL, iter_max = 10)
}
\arguments{
\item{.f}{single string like `GET`, `POST`, `PUT` and `DELETE`}

\item{...}{further named parameters which will be pass to `httr` function}

\item{.proxy}{proxy list, which has `ip` and `port` at least, if `times` was provided,
the random selection of proxy will use `times` as `prob` parameter in `sample`}

\item{.user_agent}{user agent strings vector, which is already in this package}

\item{iter_max}{the max try times}
}
\value{
list, httr response and proxy in use
}
\description{
rogue function
}
\examples{
\dontrun{
proxy <- list(
  list(ip = '1.1.1.1', port = 3128, times = 1),
  list(ip = '2.2.2.2', port = 3128, times = 2) # use times as prob in sample
)
ROGUE('GET', url = 'http://httpbin.org/get', .proxy = proxy, .user_agent = useragent)
}
}
